Add API Explorer and build app before seeding DB
Added the `AddEndpointsApiExplorer` service to the `AddPresentation` method in `WebApplicationBuilderExtensions.cs` for API endpoint discovery and documentation.

In `Program.cs`, added `var app = builder.Build();` to build the application before creating a service scope and seeding the database, ensuring the application is fully configured and ready to use the services.